壁の判定追加

+++ Data.h
@@ -9,7 +9,7 @@ struct Player_ {
 	Vector2 pos = { 32.0f,600.0f };
 	Vector2 velocity = { 0.0f,0.0f };
 	Vector2 acceleration = {0.0f,0.5f};
-	int radius = 32;
+	float radius = 32;
 	int isFlash=false;
 	int isJanp=false;
 	int janpNum =0;
